+++ pybitget/__init__.py
@@ -4,7 +4,7 @@
 
 """
 
-__version__ = '1.0.4'
+__version__ = '1.0.5'
 
 from loguru import logger
 from pybitget.client import Client