+++ release_notes.md
@@ -5,6 +5,8 @@ python wrapper with rest API, websocket API.
 
 
 # Release notes
+* Version 1.0.5 - 02 Feb 2023.
+  * Fixed bug [#5](https://github.com/cuongitl/python-bitget/issues/5)
 * Version 1.0.4 - 30 Jan 2023.
   * Full support for all api-endpoint: spot/mix/broker
   * Update some functions by [Bitget Update log](https://bitgetlimited.github.io/apidoc/en/mix/#update-log)